+++ lib/resources/actions.rb
@@ -1,13 +1,13 @@
 module Resources
   module Actions
     extend ActiveSupport::Concern
-          
+
       included do
         helper_method :resources, :resource, :resources_search
 
         if self.resource_configuration.resource_method_name
           self.class_eval <<-RUBY
-            
+
             alias_method :#{self.resource_configuration.resources_method_name.to_s}, :resources
             alias_method :#{self.resource_configuration.resource_method_name.to_s}, :resource
             alias_method :#{self.resource_configuration.resources_method_name.to_s}_search, :resources_search
@@ -26,9 +26,9 @@ module Resources
 
       def resource
         @resource ||= resource_manager.resource
-      end 
+      end
 
-      
+
 
       protected
 
@@ -36,7 +36,7 @@ module Resources
         @resource_manager ||= Resources::Manager.new(self, request)
       end
 
-      
+
 
 
       def member_route?