+++ test/dummy/app/controllers/countries_controller.rb
@@ -1,0 +1,4 @@
+class CountriesController < ApplicationController
+  resource_for "Country"
+
+end