+++ test/dummy/app/models/country.rb
@@ -1,0 +1,7 @@
+class Country < ActiveRecord::Base
+
+  def permited_attributes
+    [:name, :code, :active]
+  end
+
+end