+++ test/dummy/config/routes.rb
@@ -1,9 +1,13 @@
 Rails.application.routes.draw do
+  resources :countries
+  namespace :admin do
+    resources :countries
+  end
   # The priority is based upon order of creation: first created -> highest priority.
   # See how all your routes lay out with "rake routes".
 
   # You can have the root of your site routed with "root"
-  # root 'welcome#index'
+  root 'countries#index'
 
   # Example of regular route:
   #   get 'products/:id' => 'catalog#view'